+++ designer/client/components/LogicExpressions/logic-expression-edit.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { Select, Input } from "@xgovformbuilder/govuk-react-jsx";
 import { LogicExpressionProps } from "./types";
 import logger from "../../plugins/logger";
-import { ValidationErrors } from "../../outputs/types";
+import { ValidationErrors } from "../../outputs-visualiser/outputs/types";
 import { validateNotEmpty, hasValidationErrors } from "../../validations";
 import ErrorSummary from "../../error-summary";
