+++ designer/client/outputs-visualiser/outputs/__tests__/output-edit.jest.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, fireEvent, waitFor } from "@testing-library/react";
-import { RenderWithContextAndDataContext } from "../../__tests__/helpers/renderers";
+import { RenderWithContextAndDataContext } from "../../../__tests__/helpers/renderers";
 import { Data } from "@xgovformbuilder/model";
 
 import OutputEdit from "../output-edit";
